sysdaa/requirements.txt
asgiref==3.11.1
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
cssselect2==0.9.0
Django==6.0.3
django-environ==0.12.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.4.0
django-two-factor-auth==1.18.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.61.1
openpyxl==3.1.5
phonenumbers==9.0.25
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-dotenv==1.2.1
qrcode==8.2
reportlab==4.4.10
sqlparse==0.5.5
structlog==25.5.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.1
